refactor: :bug: replace button inside PDFDownloadLink with styled anchor
